refactor(web): one component per file, unify duplicate conditionals, use date-fns
Splits RecurringTransactionActions.tsx's four button components (plus the
shared ConfirmActionButton) into their own files under a components/
subfolder, per PR review feedback - each file/test now maps 1:1 to a single
component, matching the new AGENTS.md convention this also documents.

Also, in RecurringTransactionsSection/index.tsx:
- Collapses two adjacent blocks both gated on
  `recurrence.status !== RecurrenceStatus.CANCELLED` into one guard.
- Replaces naive `.slice(0, 10)` ISO-string slicing with `date-fns`'s
  `parseISO` + next-intl's `useFormatter().dateTime()`, matching the
  pattern already used in TransactionTableRow for locale-aware display.
