+++ AGENTS.md
@@ -34,6 +34,22 @@ Run /simplify before presenting code to the user.
 
 <!-- END:naming-conventions -->
 
+<!-- BEGIN:component-organization -->
+
+## Component Organization
+
+- One component per file. A folder with multiple related components (e.g. a set of action buttons for a
+  particular section) should have its own `components/` subfolder with one file per component, so it's clear
+  each one belongs to that parent feature.
+- Don't repeat the same condition across sibling branches, in JSX or in plain logic. If two adjacent blocks
+  guard on the same condition, wrap them in a single check (a `<>...</>` fragment for JSX) instead of
+  duplicating the check.
+- Prefer `date-fns` (already a dependency) over manual string slicing/parsing for date handling - e.g.
+  `parseISO` to turn a backend ISO string into a `Date`. For locale-aware display formatting, pair it with
+  `next-intl`'s `useFormatter().dateTime(...)`, matching the pattern already used in `TransactionTableRow`.
+
+<!-- END:component-organization -->
+
 <!-- BEGIN:nextjs-agent-rules -->
 
 # Next.js: ALWAYS read docs before coding
