+++ src/app/transactions/components/RecurringTransactionsSection/__tests__/RecurringTransactionsSection.test.tsx
@@ -14,6 +14,7 @@ import { TransactionType } from '@/types/enums/transactionType';
 jest.mock('next-intl', () => ({
   useTranslations: () => (key: string) => key,
   useLocale: () => 'en',
+  useFormatter: () => ({ dateTime: (date: Date) => date.toISOString() }),
 }));
 
 // TypeBadge imports useTranslations from 'use-intl' directly (next-intl's own re-export), which
