+++ src/app/transactions/components/RecurringTransactionsSection/components/__tests__/EditRecurringTransactionButton.test.tsx
@@ -1,0 +1,33 @@
+import { fireEvent, render, screen } from '@testing-library/react';
+import React from 'react';
+import { EditRecurringTransactionButton } from '@/app/transactions/components/RecurringTransactionsSection/components/EditRecurringTransactionButton';
+import { useTransactionForm } from '@/components/TransactionForm/TransactionFormProvider';
+import { RecurrenceStatus } from '@/types/enums/recurrenceStatus';
+import { buildRecurrence } from '../testFixtures/buildRecurrence';
+
+jest.mock('next-intl', () => ({
+  useTranslations: () => (key: string) => key,
+}));
+
+jest.mock('@/components', () => ({
+  Button: jest.requireActual('@/components/Button').Button,
+}));
+
+jest.mock('@/components/TransactionForm/TransactionFormProvider', () => ({
+  useTransactionForm: jest.fn(),
+}));
+
+const mockedUseTransactionForm = useTransactionForm as jest.Mock;
+
+describe('EditRecurringTransactionButton', () => {
+  it('opens the recurring form for the given recurrence when pressed', () => {
+    const showRecurringTransactionForm = jest.fn();
+    mockedUseTransactionForm.mockReturnValue({ showRecurringTransactionForm });
+    const activeRecurrence = buildRecurrence(RecurrenceStatus.ACTIVE);
+
+    render(<EditRecurringTransactionButton recurrence={activeRecurrence} />);
+    fireEvent.click(screen.getByRole('button', { name: /edit/i }));
+
+    expect(showRecurringTransactionForm).toHaveBeenCalledWith(activeRecurrence);
+  });
+});
